+++ service-b/src/main/java/io/x99/service_b/controller/MainController.java
@@ -1,7 +1,7 @@
-package io.x99.controller;
+package io.x99.service_b.controller;
 
-import io.x99.model.User;
-import io.x99.service.ServiceAClient;
+import io.x99.service_b.model.User;
+import io.x99.service_b.service.ServiceAClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
