+++ service-b/src/main/java/io/x99/service_b/model/ERole.java
@@ -1,0 +1,5 @@
+package io.x99.service_b.model;
+
+public enum ERole {
+
+}
