+++ service-b/src/main/java/io/x99/service_b/model/User.java
@@ -1,4 +1,4 @@
-package io.x99.model;
+package io.x99.service_b.model;
 
 public class User {
     private Long id;
