+++ service-b/src/main/java/io/x99/service_b/service/ServiceAClient.java
@@ -1,11 +1,8 @@
-package io.x99.service;
+package io.x99.service_b.service;
 
-import io.x99.model.User;
+import io.x99.service_b.model.User;
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.http.MediaType;
-import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
