+++ service-b/src/main/resources/application.properties
@@ -1,3 +1,16 @@
 server.port=8082
+
+# datasource
+spring.datasource.url= jdbc:postgresql://localhost:5432/liquibase
+spring.datasource.username= postgres
+spring.datasource.password= password
+
+spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation= true
+spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.PostgreSQLDialect
+
+# Hibernate ddl auto (create, create-drop, validate, update)
+spring.jpa.hibernate.ddl-auto= validate
+
+# Feign Client
 feign.okhttp.enabled=true
 service-a.url=http://localhost:8081
